TSDK-828 Refactor state machine
